requirements.txt
efficientnet-pytorch==0.7.0
numpy==1.20.1
Pillow==8.1.0
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
